+++ NIMIO2024_TFG_PolMorenoCastelar/src/main/java/com/nimio2024/nimio2024_tfg_polmorenocastelar/controllers/SchoolController.java
@@ -64,4 +64,10 @@ public class SchoolController {
         School school = schoolService.getSchoolById(schoolId);
         return school.getCenter().getCenterName();
     }
+
+    public School updateSchoolName(Long schoolId, String schoolName) {
+        School school = schoolService.getSchoolById(schoolId);
+        school.setSchoolName(schoolName);
+        return schoolService.saveSchool(school);
+    }
 }